kconfig: move CROSS_COMPILE_ARM32 to the correct place
 Set-performance-critical-threads-and-IRQs
 Add-cpumasks-for-big-and-LITTLE-CPU-clusters
 Add-cpu_input_boost-implementation
 Introduce-CPUSets-Assist
 Introduce-devfreq-boost-driver
 Introduce-SchedTune-Assist-v3

Signed-off-by: Carlos Jimenez (JavaShin-X) <javashin1986@gmail.com>

 arch/arm64/Kconfig                            |   14 +
 drivers/android/simple_lmk.c                  |    2 +-
 drivers/cpufreq/Kconfig                       |   37 +
 drivers/cpufreq/Makefile                      |    5 +
 drivers/cpufreq/cpu_input_boost.c             |  377 +++
 drivers/devfreq/Kconfig                       |   33 +
 drivers/devfreq/Makefile                      |    3 +
 drivers/devfreq/devfreq.c                     |   81 +-
 drivers/devfreq/devfreq_boost.c               |  332 ++
 drivers/devfreq/devfreq_devbw.c               |    4 +
 drivers/devfreq/governor.h                    |    3 -
 drivers/devfreq/governor_memlat.c             |   74 +-
 drivers/gpu/drm/drm_irq.c                     |    4 +-
 drivers/gpu/drm/msm/msm_drv.c                 |   39 +-
 include/linux/cpu_input_boost.h               |   20 +
 include/linux/cpumask.h                       |    4 +
 include/linux/devfreq.h                       |   13 +-
 include/linux/devfreq_boost.h                 |   34 +
 init/Kconfig                                  |   64 +
 kernel/cgroup/cpuset.c                        | 2832 +++++++++++++++++
 kernel/sched/tune.c                           |   57 +-
 create mode 100644 drivers/cpufreq/cpu_input_boost.c
 create mode 100644 drivers/devfreq/devfreq_boost.c
 create mode 100644 include/linux/cpu_input_boost.h
 create mode 100644 include/linux/devfreq_boost.h
 create mode 100644 kernel/cgroup/cpuset.c

(*) cpumask: Add cpumasks for big and LITTLE CPU clusters

Add cpu_lp_mask and cpu_perf_mask to represent the CPUs that belong to
each cluster in a dual-cluster, heterogeneous system.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
[kdrag0n: Forward-ported to k4.14]
Signed-off-by: Danny Lin <danny@kdrag0n.dev>
Signed-off-by: celtare21 <celtare21@gmail.com>

(*) kernel: Add API to mark IRQs and kthreads as performance critical

On devices with a CPU that contains heterogeneous cores (e.g., big.LITTLE),
it can be beneficial to place some performance-critical IRQs and kthreads
onto the performance CPU cluster in order to improve performance.

This commit adds the following APIs:
-kthread_run_perf_critical() to create and start a perf-critical kthread
-IRQF_PERF_CRITICAL to schedule an IRQ and any threads it may have onto
 performance CPUs
-PF_PERF_CRITICAL to mark a process (mainly a kthread) as performance
 critical (this is used by kthread_run_perf_critical())

In order to accommodate this new API, the following changes are made:
-Performance-critical IRQs have their affinities reaffined upon exit from
 suspend (since the affinities are broken when non-boot CPUs are disabled)
-Performance-critical IRQs and their threads have their affinities reset
 upon entering suspend, so that upon immediate suspend exit (when only the
 boot CPU is online), interrupts can be processed and interrupt threads
 can be scheduled onto an online CPU (otherwise we'd hit a kernel BUG)
-do_set_cpus_allowed() and __set_cpus_allowed_ptr() are modified to enforce
 a performance-critical kthread's affinity
-Mark perf-critical IRQs with IRQD_AFFINITY_MANAGED so userspace can't mess
 with their affinity

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>
Signed-off-by: celtare21 <celtare21@gmail.com>

(*) devfreq_boost: Introduce devfreq boost driver

This driver boosts enumerated devfreq devices upon input, and allows for
boosting specific devfreq devices on other custom events. The boost
frequencies for this driver should be set so that frame drops are
near-zero at the boosted frequencies and power consumption is minimized
at said frequencies. The goal of this driver is to provide an interface
to achieve optimal device performance by requesting boosts on key
events, such as when a frame is ready to rendered to the display.

Currently, support is only present for boosting the cpubw devfreq
device, but the driver is structured in a way that makes it easy to add
support for new boostable devfreq devices in the future.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
[yaro: adjusted for 4.14 (soc:qcom,cpubw->soc:qcom,cpu-llcc-ddr-bw]
Signed-off-by: Yaroslav Furman <yaro330@gmail.com>
Signed-off-by: Panchajanya1999 <rsk52959@gmail.com>
Signed-off-by: Panchajanya1999 <panchajanya@azure-dev.live>

(*) cpu_input_boost: Introduce driver for event-based CPU boosting

This is a simple CPU input boost driver that boosts all online CPUs for
a fixed amount of time. Additionally, there is an API for other drivers
to request a boost kick (or a max-boost kick), so boosting can be done
on any custom event. This API is mainly intended for the framebuffer
driver to send a boost kick whenever there is a new frame ready to be
rendered to the display.

This driver also boosts all online CPUs to their maximum frequencies
when the display is powered on (this is the wake boost).

Since this driver requires careful tuning for optimal performance, there
are no user-exposed knobs to configure it. All necessary configuration
is done via the supplied Kconfig options.

This driver is designed for heterogeneous multi-processor systems with
two CPU clusters.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Yaroslav Furman <yaro330@gmail.com>